# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API service for chat functionality",
        "title": "Notify Chat Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/chats/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new chat message (channel or direct)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create a new chat message",
                "parameters": [
                    {
                        "description": "Chat message data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_models.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Chat message created",
                        "schema": {
                            "$ref": "#/definitions/chat-service_internal_models.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/messages/channel/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all chat messages for a specific channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Get chat messages in a channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chat messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chat-service_internal_models.ChatResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Channel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chat-service_internal_models.ChatRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "channelId": {
                    "description": "for channel",
                    "type": "integer"
                },
                "fileName": {
                    "type": "string"
                },
                "receiverId": {
                    "description": "for direct",
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "direct",
                        "channel"
                    ]
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "chat-service_internal_models.ChatResponse": {
            "type": "object",
            "properties": {
                "channelId": {
                    "description": "channel",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "receiverId": {
                    "description": "Relate to type message",
                    "type": "integer"
                },
                "senderId": {
                    "type": "integer"
                },
                "senderName": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "description": "\"direct\" | \"channel\"",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}